# speedup testsuite