--- a/.luacheckrc
+++ b/.luacheckrc
@@ -28,6 +28,8 @@ files["spec/unit/*"].std = "+busted"
 files["spec/unit/*"].globals = {
     "match", -- can be removed once luacheck 0.24.0 or higher is used
     "package",
+    "disable_plugins",
+    "load_plugin",
 }
 
 -- TODO: clean up and enforce max line width (631)