--- a/spec/unit/autosuspend_spec.lua
+++ b/spec/unit/autosuspend_spec.lua
@@ -1,7 +1,6 @@
 describe("AutoSuspend", function()
     setup(function()
         require("commonrequire")
-        package.unloadAll()
         require("document/canvascontext"):init(require("device"))
     end)