--- a/spec/unit/autosuspend_spec.lua
+++ b/spec/unit/autosuspend_spec.lua
@@ -1,6 +1,7 @@
 describe("AutoSuspend", function()
     setup(function()
         require("commonrequire")
+        disable_plugins()
         require("document/canvascontext"):init(require("device"))
     end)