--- a/spec/unit/device_spec.lua
+++ b/spec/unit/device_spec.lua
@@ -31,7 +31,6 @@ describe("device module", function()
             end
         }
         require("commonrequire")
-        package.unloadAll()
         ffi = require("ffi")
         C = ffi.C
         require("ffi/linux_input_h")