--- a/spec/unit/device_spec.lua
+++ b/spec/unit/device_spec.lua
@@ -31,6 +31,7 @@ describe("device module", function()
             end
         }
         require("commonrequire")
+        disable_plugins()
         ffi = require("ffi")
         C = ffi.C
         require("ffi/linux_input_h")