--- a/spec/unit/filemanager_spec.lua
+++ b/spec/unit/filemanager_spec.lua
@@ -2,7 +2,6 @@ describe("FileManager module", function()
     local DataStorage, FileManager, lfs, docsettings, UIManager, Screen, makePath, util
     setup(function()
         require("commonrequire")
-        package.unloadAll()
         require("document/canvascontext"):init(require("device"))
         DataStorage = require("datastorage")
         FileManager = require("apps/filemanager/filemanager")