--- a/spec/unit/readerlink_spec.lua
+++ b/spec/unit/readerlink_spec.lua
@@ -3,7 +3,6 @@ describe("ReaderLink module", function()
 
     setup(function()
         require("commonrequire")
-        package.unloadAll()
         require("document/canvascontext"):init(require("device"))
         DocumentRegistry = require("document/documentregistry")
         Event = require("ui/event")