… spec/unit/readerhighlight_spec.lua
… spec/unit/readerpaging_spec.lua
… spec/unit/readerrolling_spec.lua
… spec/unit/readersearch_spec.lua
… spec/unit/readertoc_spec.lua
… spec/unit/readerui_spec.lua
… spec/unit/readerview_spec.lua
… spec/unit/screenshoter_spec.lua
… spec/unit/translator_spec.lua
… spec/unit/wakeupmgr_spec.lua