- drop unnecessary workarounds
- cleanup & speedup tests
- fix screenshots
- fix instance mismatch errors

Timings on my machine (seconds):

|                                   | master|   PR        |
------------------------------------|------:|------------:|
`./kodev test -bnj1` |  42.6 | 20.2 (÷2.1) |
`./kodev test -bnj8` |  10.4 |  5.2 (÷2.0) |


<!-- Reviewable:start -->
- - -
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/koreader/koreader/12906)
<!-- Reviewable:end -->